--- src/App.tsx
@@ -1,12 +1,0 @@
-import Homepage from './components/Homepage';
-import './App.css';
-
-function App() {
-  return (
-    <div className="App bg-[#282828] text-[#F1F1F1] pt-2">
-      {/* <Homepage/> */}
-    </div>
-  );
-}
-
-export default App;
