+++ src/components/Browser.tsx
@@ -41,7 +41,7 @@ export type data = {
   [key: string]: fileData | folderData
 }
 
-function Homepage({route}: {route: string}) {
+function Browser({route}: {route: string}) {
   return (
     <>
       <Header route={route}/>
@@ -50,4 +50,4 @@ function Homepage({route}: {route: string}) {
   );
 }
 
-export default Homepage;
+export default Browser;
