+++ src/components/GridView.tsx
@@ -3,7 +3,7 @@ import { createTheme, ThemeProvider }from '@mui/material/styles';
 import Grid from '@mui/material/Grid';
 import Box from '@mui/material/Box';
 import Dirent from './Dirent';
-import {data} from "./Homepage";
+import { data } from "./Browser";
 
 declare module '@mui/material/styles' {
   interface BreakpointOverrides {
@@ -42,18 +42,16 @@ function GridView({dataObj, route}: {dataObj: data, route: string}) {
               s1={12} s2={6} s3={4} s4={3}
               s5={2.4} s6={2} s7={1.7143} s8={1.5}
               s9={1.3333} s10={1.2} s11={1.0909} s12={1}
-              onClick={(event) => {
+              onClick={_ => {
                 clicks++;
                 if (clicks == 1) {
-                  setTimeout(function(){
-                    if(clicks == 1) {
-                      entry += '/:review' 
-                      router.push(entry);
-                    } else {
-                      router.push(entry);
-                    }
+                  setTimeout(function() {
+                    if(clicks == 1)
+                      router.push('/review' + entry);
+                    else
+                      router.push('/browse' + entry);
                     clicks = 0;
-                  }, 300);
+                  }, 200);
                 }
               }}
             >
@@ -66,19 +64,14 @@ function GridView({dataObj, route}: {dataObj: data, route: string}) {
               s1={12} s2={6} s3={4} s4={3}
               s5={2.4} s6={2} s7={1.7143} s8={1.5}
               s9={1.3333} s10={1.2} s11={1.0909} s12={1}
-              onClick={(event) => {
+              onClick={_ => {
                 clicks++;
                 if (clicks == 1) {
                   setTimeout(function(){
-                    if(clicks == 1) {
-                      entry += '/:review' 
-                      router.push(entry);
-                    } else {
-                      console.log('Double Clicked File')
-                      // router.push(entry);
-                    }
+                    if(clicks == 1)
+                      router.push('/review' + entry);
                     clicks = 0;
-                  }, 300);
+                  }, 200);
                 }
               }}
             >
